requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pydantic==2.10.4
google-search-results==2.4.2
requests==2.31.0
python-multipart==0.0.6
httpx==0.26.0
asyncpg==0.29.0
psycopg-pool==3.2.0
pytest==7.4.4
